+++ RELEASE_NOTES.md
@@ -21,7 +21,6 @@ Not yet funded, but would be really nice to have:
 
 Funded, but not yet done:
  * geoserver: update to 2.5.x with geofence - see [#677](https://github.com/georchestra/georchestra/issues/677) - CIGALsace & GeoBretagne,
- * mapfishapp: export current map context as a "map metadata" - see [#751](https://github.com/georchestra/georchestra/issues/751) - GéoPicardie,
  * mapfishapp: support read/write WMTS layer types from/to OWSContext - see [#360](https://github.com/georchestra/georchestra/issues/360) & [#399](https://github.com/georchestra/georchestra/issues/399) - CIGALsace,
  * mapfishapp: revamp of the "add layers" window - see [#718](https://github.com/georchestra/georchestra/issues/718) - CIGALsace,
  * ldapadmin: delegated administrators - see [#786](https://github.com/georchestra/georchestra/issues/786) - GeoBretagne,
@@ -38,6 +37,7 @@ New features:
  * mapfishapp: png printing - see [#814](https://github.com/georchestra/georchestra/issues/814),
  * mapfishapp: print extent handling improvement - see [#813](https://github.com/georchestra/georchestra/issues/813),
  * mapfishapp: give easy access to WMS/WFS layer details - see [#346](https://github.com/georchestra/georchestra/issues/346),
+ * mapfishapp: export current map context as a "map metadata" - see [#751](https://github.com/georchestra/georchestra/issues/751),
 
 Enhancements:
  * extractorapp: editable combobox for resolution selection - see [#726](https://github.com/georchestra/georchestra/issues/726),